+++ src/Components/index.js
@@ -2,8 +2,5 @@ export * from "./card";
 export * from "./modal";
 export * from "./sideNav";
 export * from "./header";
-export * from "./mediaComp";
 export * from "./tabs";
-export * from "./typeComp";
 export * from "./conditionalRenderer";
-export * from "./sideContent";
