+++ src/Components/tabs/styles.scss
@@ -43,16 +43,15 @@
         font-size: 1rem;
         font-weight: 500;
         margin: 0px;
-        padding-top: 20px;
       }
       .inner-text:not(:first-child) {
         padding-left: 10px;
       }
       .progress-bar {
         .ant-progress-inner {
-          width: 80px !important;
-          height: 80px !important;
-          font-size: 16px !important;
+          width: 50px !important;
+          height: 50px !important;
+          font-size: 12px !important;
         }
       }
     }
