+++ src/Pages/details/details.scss
@@ -9,12 +9,12 @@
 }
 .header {
   width: 100%;
-  background: linear-gradient(
-    to right,
-    rgba(31.5, 31.5, 31.5, 1) calc((50vw - 170px) - 340px),
-    rgba(31.5, 31.5, 31.5, 0.84) 50%,
-    rgba(31.5, 31.5, 31.5, 0.84) 100%
-  ) !important;
+  // background: linear-gradient(
+  //   to right,
+  //   rgba(31.5, 31.5, 31.5, 1) calc((50vw - 170px) - 340px),
+  //   rgba(31.5, 31.5, 31.5, 0.84) 50%,
+  //   rgba(31.5, 31.5, 31.5, 0.84) 100%
+  // ) !important;
 }
 :where(.css-dev-only-do-not-override-yp8pcc).ant-modal .ant-modal-content {
   background-color: black !important;
