+++ src/Pages/details/header/styles.scss
@@ -1,0 +1,11 @@
+.genres-btn {
+  background: linear-gradient(
+    to right,
+    rgb(230, 226, 226),
+    rgba(255, 122, 89, 0)
+  );
+
+  color: white;
+  font-weight: 700;
+  margin-left: 10px;
+}
