+++ src/Pages/index.js
@@ -1,5 +1,5 @@
 export * from "./dashboard";
 export * from "./details";
 export * from "./logout";
-export * from "./sign-up/SignUp";
+export * from "./sign-up/signUp";
 export * from "./logout";
