--- src/Pages/tv/dashboard.scss
@@ -1,44 +1,0 @@
-.dashboard-Container {
-  padding-left: 15px;
-  padding-right: 15px;
-  .mainDivHeader {
-    margin-top: 100px;
-    padding-left: 40px;
-    padding-right: 80px;
-    .mainTitleHeading {
-      margin: 0px;
-      color: black;
-      padding-bottom: 50px;
-    }
-  }
-  .header-img-div {
-    width: 100%;
-    max-height: 100%;
-    min-height: 100%;
-    display: flex;
-    justify-content: flex-end;
-    .header-img {
-      width: 100%;
-      position: absolute;
-      min-height: 100%;
-      max-height: 100%;
-      border-radius: 2%;
-    }
-  }
-  .select-category {
-    display: flex;
-    justify-content: flex-end;
-    padding-top: 20px;
-  }
-  .card-main-div {
-    padding-top: 30px;
-    width: 100%;
-    display: inline-flex;
-    height: 38vh;
-    overflow: visible;
-    overflow-y: hidden;
-    position: relative;
-    z-index: 1;
-  }
-  
-}
