+++ src/packages/index.js
@@ -1,0 +1,6 @@
+import { Images } from "./images";
+import { Videos } from "./videos";
+import { MovieSideDetail } from "./movieSideDetail";
+import { MoviesList } from "./moviesList";
+
+export { Images, Videos, MovieSideDetail, MoviesList };
