+++ src/packages/videos/styles.scss
@@ -1,0 +1,9 @@
+.videos-main-div{
+    width: 100%;
+    display: inline-flex;
+    height: 45vh;
+    overflow: auto;
+    overflow-y: hidden;
+    position: relative;
+    z-index: 1;
+  }
