+++ src/utils/ApiHandler.js
@@ -1,20 +1,16 @@
 import axios from "axios";
-import { api_key, moviesdata, tokenURL } from "../Api's URL/apiUrl";
+import { api_key, moviesdata, tokenURL } from "./Api's URL/apiUrl";
 const ApiHandler = async (method, url) => {
   if (method === "GET") {
     try {
-      const response = await axios.get(
-        `${tokenURL}?api_key=${api_key}`
-      );
+      const response = await axios.get(`${tokenURL}?api_key=${api_key}`);
       if (response) {
         return response;
       }
     } catch (error) {}
   } else if (method === "moviesData") {
     try {
-      const response = await axios.get(
-        `${moviesdata}=${url}`
-      );
+      const response = await axios.get(`${moviesdata}=${url}`);
       if (response) {
         return response;
       }
