+++ src/utils/index.js
@@ -1,0 +1,4 @@
+import useFetchHook from "./useFetch";
+import useLazyFetch from "./useLazyFetch";
+
+export { useFetchHook, useLazyFetch };
